move smt delta calculation to ProverStdIO for thread safety
