+++ src/main/scala/decider/ProverStdIO.scala
@@ -40,6 +40,7 @@ abstract class ProverStdIO(uniqueId: String,
 
   var proverPath: Path = _
   var lastModel : String = _
+  var lastStatistics: Map[String, String] = Map.empty
 
   def name: String
   def minVersion: Version
@@ -319,7 +320,7 @@ abstract class ProverStdIO(uniqueId: String,
     }
   }
 
-  def statistics(): Map[String, String] = {
+  private def queryStatistics(): Map[String, String] = {
     var repeat = true
     var line = ""
     var stats = scala.collection.immutable.SortedMap[String, String]()
@@ -345,6 +346,45 @@ abstract class ProverStdIO(uniqueId: String,
     } while (repeat)
 
     toMap(stats)
+  }
+
+  def statistics(): Map[String, String] = {
+    lastStatistics = queryStatistics()
+    lastStatistics
+  }
+
+  /**
+    * Queries the statistics, and calculates the diff between it and the last statistics query.
+    * The difference is calculated if value can be converted to an int or double
+    * @return map with the current statistics, and the differences (only containing values that could be converted)
+    *         and keys with appended "-delta"
+    */
+  def deltaStatistics(): Map[String, String] = {
+    val currentStatistics = queryStatistics()
+    val deltaStatistics = currentStatistics map getDelta filter { case (_, value) => value.nonEmpty } map {
+      case (key, Some(value)) => (key + "-delta", value)
+      case other => sys.error(s"Unexpected result pair $other")
+    }
+    lastStatistics = lastStatistics ++ currentStatistics
+    currentStatistics ++ deltaStatistics
+  }
+
+  private def getDelta(pair: (String, String)): (String, Option[String]) = {
+    val curValInt = pair._2.toIntOption
+    val prevValInt = lastStatistics.get(pair._1) match {
+      case Some(value) => value.toIntOption
+      case _ => Some(0) // value not found
+    }
+    val curValDouble = pair._2.toDoubleOption
+    val prevValDouble = lastStatistics.get(pair._1) match {
+      case Some(value) => value.toDoubleOption
+      case _ => Some(0.0) // value not found
+    }
+    (curValInt, prevValInt, curValDouble, prevValDouble) match {
+      case (Some(curInt), Some(prevInt), _, _) => (pair._1, Some((curInt - prevInt).toString))
+      case (_, _, Some(curDouble), Some(prevDouble)) => (pair._1, Some((curDouble - prevDouble).toString))
+      case _ => (pair._1, None)
+    }
   }
 
   def comment(str: String): Unit = {
